refactor: add restaurant.ts
