+++ src/library/map/types.ts
@@ -34,10 +34,3 @@ export interface RawResult {
 	x: string;
 	y: string;
 }
-
-export interface Restaurant extends SearchResult {
-	[index: string]: any;
-	star: number;
-	review: number;
-	color: string;
-}
